+++ src/main/kotlin/shop/itbug/fluttercheckversionx/common/MyAction.kt
@@ -3,7 +3,6 @@ package shop.itbug.fluttercheckversionx.common
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.ToggleAction
-import com.intellij.openapi.actionSystem.ex.ComboBoxAction
 import com.intellij.openapi.project.DumbAwareAction
 import java.util.function.Supplier
 import javax.swing.Icon
@@ -28,8 +27,6 @@ abstract class MyDumbAwareAction : DumbAwareAction {
     constructor(name: Supplier<String>) : super(name)
     constructor(icon: Icon) : super(icon)
 
-    constructor(title: String, desc: String, icon: Icon) : super(title, desc, icon)
-
     override fun getActionUpdateThread(): ActionUpdateThread {
         return ActionUpdateThread.BGT
     }
@@ -43,9 +40,3 @@ abstract class MyToggleAction(name: Supplier<String>) : ToggleAction(name) {
     }
 }
 
-abstract class MyComboBoxAction : ComboBoxAction() {
-
-    override fun getActionUpdateThread(): ActionUpdateThread {
-        return ActionUpdateThread.BGT
-    }
-}