+++ src/main/kotlin/shop/itbug/fluttercheckversionx/notif/PubPluginVersionCheckNotification.kt
@@ -1,12 +1,10 @@
 package shop.itbug.fluttercheckversionx.notif
 
 import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.fileEditor.FileEditor
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.ui.EditorNotificationPanel
 import com.intellij.ui.EditorNotificationProvider
 import com.intellij.ui.EditorNotifications
@@ -18,15 +16,15 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import org.jetbrains.yaml.psi.YAMLFile
 import shop.itbug.fluttercheckversionx.dialog.SearchDialog
-import shop.itbug.fluttercheckversionx.editor.MyDartPackageTree
 import shop.itbug.fluttercheckversionx.i18n.PluginBundle
 import shop.itbug.fluttercheckversionx.icons.MyIcons
 import shop.itbug.fluttercheckversionx.services.DartPackageCheckService
+import shop.itbug.fluttercheckversionx.services.noused.DartNoUsedCheckService
 import shop.itbug.fluttercheckversionx.setting.IgPluginPubspecConfigList
 import shop.itbug.fluttercheckversionx.tools.MyToolWindowTools
-import shop.itbug.fluttercheckversionx.tools.showInCenterOfPopup
-import shop.itbug.fluttercheckversionx.util.MyPsiElementUtil
+import shop.itbug.fluttercheckversionx.util.MyFileUtil
 import shop.itbug.fluttercheckversionx.util.getPubspecYAMLFile
+import shop.itbug.fluttercheckversionx.widget.DartPackageTable
 import java.util.function.Function
 import javax.swing.JComponent
 
@@ -45,7 +43,6 @@ class PubPluginVersionCheckNotification : EditorNotificationProvider, DumbAware 
             pubFile =
                 ApplicationManager.getApplication().executeOnPooledThread(Callable { project.getPubspecYAMLFile() })
                     .get()
-            println("file is $pubFile")
             if (it.component.parent != null && pubFile != null) {
                 EditorNotifications.getInstance(project).updateNotifications(pubFile!!.virtualFile)
             }
@@ -61,7 +58,8 @@ class YamlFileNotificationPanel(fileEditor: FileEditor, val project: Project) :
     EditorNotificationPanel(fileEditor, UIUtil.getEditorPaneBackground()) {
 
     private var checkLabel: HyperlinkLabel = createActionLabel(PluginBundle.get("check.flutter.plugin")) {
-        MyDartPackageTree.createPanel(project).showInCenterOfPopup(project)
+//        MyDartPackageTree.createPanel(project).showInCenterOfPopup(project)
+        DartPackageTable(project).show()
     }
 
     init {
@@ -78,8 +76,8 @@ class YamlFileNotificationPanel(fileEditor: FileEditor, val project: Project) :
 
         ///重新索引
         val reIndexLabel = createActionLabel(PluginBundle.get("pubspec_yaml_file_re_index")) {
-            doReIndex()
             GlobalScope.launch { DartPackageCheckService.getInstance(project).resetIndex() }
+            MyFileUtil.reIndexPubspecFile(project)
         }
         myLinksPanel.add(reIndexLabel)
 
@@ -98,6 +96,12 @@ class YamlFileNotificationPanel(fileEditor: FileEditor, val project: Project) :
             IgPluginPubspecConfigList.showInPopup(project)
         }
         myLinksPanel.add(igPackageLabel)
+
+        ///检查没有被使用的包
+        val noUsedCheck = createActionLabel(PluginBundle.get("check_un_used_package")) {
+            DartNoUsedCheckService.getInstance(project).checkUnUsedPackaged()
+        }
+        myLinksPanel.add(noUsedCheck)
     }
 
     ///打开隐私扫描工具窗口
@@ -111,21 +115,6 @@ class YamlFileNotificationPanel(fileEditor: FileEditor, val project: Project) :
                 }
             }
         }
-    }
-
-    ///重新索引
-    @OptIn(DelicateCoroutinesApi::class)
-    private fun doReIndex() {
-        GlobalScope.launch {
-            MyPsiElementUtil.getPubSpecYamlFile(project)?.let { _ ->
-                run {
-                    WriteCommandAction.runWriteCommandAction(project) {
-                        VirtualFileManager.getInstance().refreshWithoutFileWatcher(false)
-                    }
-                }
-            }
-        }
-
     }
 
     private fun search() {